Reverted changes which added the visibility attributes.
Thanks @SpaceIm for the PR. However, these are two in one.
The one is the macro which we agree that this should be
made IIR-specific and also that templates cannot be exported
so thank you very much. Well spotted!

Regarding linker attributes:

https://stackoverflow.com/questions/52719364/how-to-use-the-attribute-visibilitydefault
https://stackoverflow.com/questions/71284410/what-exactly-setcmake-cxx-visibility-preset-hidden-does-with-cmake

attributes are really just meant to fix things
very close to the compilter/hardware such as packed structs
but they are not a general purpose language element.

So, sorry but that's not merged.